Options cleanup (token validation)
