+++ source/helpers/used-storage.ts
@@ -1,12 +1,14 @@
 /** `getBytesInUse` polyfill */
 export default async function getStorageBytesInUse(area: 'local' | 'sync'): Promise<any> {
 	const storage = chrome.storage[area];
-	return 'getBytesInUse' in storage
-		? storage.getBytesInUse() // Good browsers
-		: new TextEncoder().encode( // Firefox
-			// @ts-expect-error https://bugzilla.mozilla.org/show_bug.cgi?id=1385832#c20
+	try {
+		return await storage.getBytesInUse(); // Exists in Safari iOS, but can't be called...
+	} catch {
+		// Firefox https://bugzilla.mozilla.org/show_bug.cgi?id=1385832#c20
+		return new TextEncoder().encode( 
 			Object.entries(await storage.get())
 				.map(([key, value]) => key + JSON.stringify(value))
 				.join(''),
 		).length;
+	}
 }
