+++ source/options.tsx
@@ -3,7 +3,6 @@ import './options.css';
 import {expectElement as $, $$} from 'select-dom';
 import fitTextarea from 'fit-textarea';
 import prettyBytes from 'pretty-bytes';
-import {assertError} from 'ts-extras';
 import {enableTabToIndent} from 'indent-textarea';
 import delegate, {DelegateEvent} from 'delegate-it';
 import {isChrome, isFirefox} from 'webext-detect';
@@ -17,92 +16,13 @@ import {perDomainOptions} from './options-storage.js';
 import isDevelopmentVersion from './helpers/is-development-version.js';
 import {doesBrowserActionOpenOptions} from './helpers/feature-utils.js';
 import {state as bisectState} from './helpers/bisect.js';
-import {parseTokenScopes} from './github-helpers/github-token.js';
 import {scrollIntoViewIfNeeded} from './github-helpers/index.js';
 import initFeatureList, {updateListDom} from './options/feature-list.js';
-
-type TokenType = 'classic' | 'fine_grained';
+import initTokenValidation from './options/token-validation.js';
 
 let syncedForm: SyncedForm | undefined;
 
-type Status = {
-	tokenType: TokenType;
-	error?: true;
-	text?: string;
-	scopes?: string[];
-};
-
 const {version} = chrome.runtime.getManifest();
-
-function reportStatus({tokenType, error, text, scopes}: Status): void {
-	const tokenStatus = $('#validation');
-	tokenStatus.textContent = text ?? '';
-	if (error) {
-		tokenStatus.dataset.validation = 'invalid';
-	} else {
-		delete tokenStatus.dataset.validation;
-	}
-
-	// Toggle the ulists by token type (default to classic)
-	for (const ulist of $$('[data-token-type]')) {
-		ulist.style.display = ulist.dataset.tokenType === tokenType ? '' : 'none';
-	}
-
-	for (const scope of $$('[data-scope]')) {
-		if (scopes) {
-			scope.dataset.validation = scopes.includes(scope.dataset.scope!) ? 'valid' : 'invalid';
-		} else {
-			scope.dataset.validation = '';
-		}
-	}
-}
-
-function getApiUrl(): string {
-	const tokenLink = $('a#personal-token-link');
-	return tokenLink.host === 'github.com'
-		? 'https://api.github.com'
-		: `${tokenLink.origin}/api/v3`;
-}
-
-async function getNameFromToken(token: string): Promise<string> {
-	const response = await fetch(
-		getApiUrl() + '/user',
-		{
-			headers: {
-				Authorization: `Bearer ${token}`,
-			},
-		},
-	);
-
-	const details = await response.json();
-	if (!response.ok) {
-		throw new Error(details.message);
-	}
-
-	return details.login;
-}
-
-async function getTokenScopes(personalToken: string): Promise<string[]> {
-	const response = await fetch(getApiUrl(), {
-		cache: 'no-store',
-		headers: {
-			'User-Agent': 'Refined GitHub',
-			'Accept': 'application/vnd.github.v3+json',
-			'Authorization': `token ${personalToken}`,
-		},
-	});
-
-	if (!response.ok) {
-		const details = await response.json();
-		throw new Error(details.message);
-	}
-
-	return parseTokenScopes(response.headers);
-}
-
-function expandTokenSection(): void {
-	$('details#token').open = true;
-}
 
 async function updateStorageUsage(area: 'sync' | 'local'): Promise<void> {
 	const storage = chrome.storage[area];
@@ -114,38 +34,6 @@ async function updateStorageUsage(area: 'sync' | 'local'): Promise<void> {
 			: available < 100_000
 				? `Only ${prettyBytes(available)} available`
 				: `${prettyBytes(used)} used`;
-	}
-}
-
-async function validateToken(): Promise<void> {
-	const tokenField = $('input[name="personalToken"]');
-	const tokenType = tokenField.value.startsWith('github_pat_') ? 'fine_grained' : 'classic';
-	reportStatus({tokenType});
-
-	if (!tokenField.validity.valid || tokenField.value.length === 0) {
-	// The Chrome options iframe auto-sizer causes the "scrollIntoView" function to scroll incorrectly unless you wait a bit
-	// https://github.com/refined-github/refined-github/issues/6807
-		setTimeout(expandTokenSection, 100);
-		return;
-	}
-
-	reportStatus({text: 'Validating…', tokenType});
-
-	try {
-		const [scopes, user] = await Promise.all([
-			getTokenScopes(tokenField.value),
-			getNameFromToken(tokenField.value),
-		]);
-		reportStatus({
-			tokenType,
-			text: `👤 @${user}`,
-			scopes,
-		});
-	} catch (error) {
-		assertError(error);
-		reportStatus({tokenType, error: true, text: error.message});
-		expandTokenSection();
-		throw error;
 	}
 }
 
@@ -237,7 +125,7 @@ async function generateDom(): Promise<void> {
 	$('#js-failed').remove();
 
 	// Enable token validation
-	void validateToken();
+	void initTokenValidation(syncedForm);
 
 	// Update rate link if necessary
 	updateRateLink();
@@ -259,15 +147,9 @@ async function generateDom(): Promise<void> {
 
 function addEventListeners(): void {
 	// Update domain-dependent page content when the domain is changed
-	syncedForm?.onChange(async domain => {
-		$('a#personal-token-link').host = domain === 'default' ? 'github.com' : domain;
+	syncedForm?.onChange(async () => {
 		// Delay to let options load first
-		setTimeout(() => {
-			validateToken();
-
-			// Re-sort list
-			updateListDom();
-		}, 100);
+		setTimeout(updateListDom, 100);
 	});
 
 	// Refresh page when permissions are changed (because the dropdown selector needs to be regenerated)
@@ -300,9 +182,6 @@ function addEventListeners(): void {
 	$('#toggle-all-features').addEventListener('click', enableToggleAll);
 	$('#disable-all-features').addEventListener('click', disableAllFeatures);
 	$('#enable-all-features').addEventListener('click', enableAllFeatures);
-
-	// Add token validation
-	$('[name="personalToken"]').addEventListener('input', validateToken);
 }
 
 async function init(): Promise<void> {
